Add event participation

+++ src/main/java/com/example/user_microservice/dto/event/EventDto.java
@@ -36,7 +36,7 @@ public class EventDto {
     @NotBlank
     private String location;
 
-    private Integer maxAttendees;
+    private Long maxAttendees;
 
     @NotEmpty
     private List<Long> skills;
